Addressed feedback from code review.
