+++ FlowMagic-SDK.xcodeproj/project.xcworkspace/contents.xcworkspacedata
