+++ FlowMagic-SDK.xcodeproj/xcshareddata/xcschemes/FlowMagic-SDK-Package.xcscheme
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <Scheme
-   LastUpgradeVersion = "1410"
+   LastUpgradeVersion = "9999"
    version = "1.3">
    <BuildAction
       parallelizeBuildables = "YES"
@@ -14,8 +14,8 @@
             buildForAnalyzing = "YES">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "97F12D2629B44EAD009F74B3"
-               BuildableName = "FlowMagic-SDK.app"
+               BlueprintIdentifier = "flowmagic-sdk::FlowMagic-SDK"
+               BuildableName = "FlowMagic_SDK.framework"
                BlueprintName = "FlowMagic-SDK"
                ReferencedContainer = "container:FlowMagic-SDK.xcodeproj">
             </BuildableReference>
@@ -30,11 +30,10 @@
       codeCoverageEnabled = "YES">
       <Testables>
          <TestableReference
-            skipped = "NO"
-            parallelizable = "YES">
+            skipped = "NO">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "97F12D7429B56839009F74B3"
+               BlueprintIdentifier = "flowmagic-sdk::FlowMagic-SDKTests"
                BuildableName = "FlowMagic-SDKTests.xctest"
                BlueprintName = "FlowMagic-SDKTests"
                ReferencedContainer = "container:FlowMagic-SDK.xcodeproj">
@@ -52,16 +51,6 @@
       debugDocumentVersioning = "YES"
       debugServiceExtension = "internal"
       allowLocationSimulation = "YES">
-      <BuildableProductRunnable
-         runnableDebuggingMode = "0">
-         <BuildableReference
-            BuildableIdentifier = "primary"
-            BlueprintIdentifier = "97F12D2629B44EAD009F74B3"
-            BuildableName = "FlowMagic-SDK.app"
-            BlueprintName = "FlowMagic-SDK"
-            ReferencedContainer = "container:FlowMagic-SDK.xcodeproj">
-         </BuildableReference>
-      </BuildableProductRunnable>
    </LaunchAction>
    <ProfileAction
       buildConfiguration = "Release"
@@ -69,16 +58,6 @@
       savedToolIdentifier = ""
       useCustomWorkingDirectory = "NO"
       debugDocumentVersioning = "YES">
-      <BuildableProductRunnable
-         runnableDebuggingMode = "0">
-         <BuildableReference
-            BuildableIdentifier = "primary"
-            BlueprintIdentifier = "97F12D2629B44EAD009F74B3"
-            BuildableName = "FlowMagic-SDK.app"
-            BlueprintName = "FlowMagic-SDK"
-            ReferencedContainer = "container:FlowMagic-SDK.xcodeproj">
-         </BuildableReference>
-      </BuildableProductRunnable>
    </ProfileAction>
    <AnalyzeAction
       buildConfiguration = "Debug">
