+++ Package.swift
@@ -5,6 +5,9 @@ import PackageDescription
 
 let package = Package(
     name: "FlowMagic-SDK",
+    platforms: [
+        .macOS(.v12), .iOS(.v15)
+    ],
     products: [
         // Products define the executables and libraries a package produces, and make them visible to other packages.
         .library(
