+++ Sources/FlowMagic-SDK/Assets.xcassets/AccentColor.colorset/Contents.json
