+++ Sources/FlowMagic-SDK/Assets.xcassets/AppIcon.appiconset/Contents.json
