+++ Sources/FlowMagic-SDK/Assets.xcassets/Contents.json
