+++ Sources/FlowMagic-SDK/Model/screenFlowModel.swift
