+++ Sources/FlowMagic-SDK/Utils/Constants.swift
