+++ Sources/FlowMagic-SDK/ViewModels/flowMagicViewModel.swift
