--- Tests/FlowMagic-SDKTests/FlowMagic_SDKTests.swift
@@ -1,11 +1,0 @@
-import XCTest
-@testable import FlowMagic_SDK
-
-final class FlowMagicSDKTests: XCTestCase {
-    func testExample() throws {
-        // This is an example of a functional test case.
-        // Use XCTAssert and related functions to verify your tests produce the correct
-        // results.
-//        XCTAssertEqual(FlowMagic_SDK().text, "Hello, World!")
-    }
-}
